refactor(web): standardize file naming from PascalCase to kebab-case
- Rename RequireAuth.tsx → require-auth.tsx
- Rename RequireRole.tsx → require-role.tsx
- Rename NotificationProvider.tsx → notification-provider.tsx
- Update all import paths to match new file names
- Fix missing 't' dependency in useCallback hook

This refactoring improves consistency with common React/Next.js conventions
where component files use kebab-case naming.
